pyproject.toml
[project]
name = "ninjserv-daily-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.5.2",
    "python-dotenv>=1.1.0",
    "motor>=3.3.2",
    "numpy>=1.24.0",
    "pymongo>=4.6.0",
    "google-genai>=1.27.0",
    "pydantic>=2.11.7",
]
